Fix PyType errors
