+++ skyplane/__init__.py
@@ -3,7 +3,8 @@ from pathlib import Path
 from skyplane import compute, exceptions
 from skyplane.api.client import SkyplaneClient
 from skyplane.api.config import TransferConfig, AWSConfig, AzureConfig, GCPConfig
-from skyplane.api.provision.dataplane import Dataplane
+from skyplane.api.dataplane import Dataplane
+from skyplane.api.tracker import TransferHook
 
 __version__ = "0.2.1"
 __root__ = Path(__file__).parent.parent
@@ -20,4 +21,5 @@ __all__ = [
     "AWSConfig",
     "AzureConfig",
     "GCPConfig",
+    "TransferHook",
 ]
