+++ skyplane/api/dataplane.py
@@ -10,17 +10,16 @@ import urllib3
 from typing import TYPE_CHECKING, Dict, List, Optional
 
 from skyplane import compute
-from skyplane.api.tracker import TransferProgressTracker
+from skyplane.api.tracker import TransferProgressTracker, TransferHook
 from skyplane.api.transfer_job import CopyJob, SyncJob, TransferJob
 from skyplane.api.config import TransferConfig
-from skyplane.progress_reporting.transfer_hooks import TransferHook
 from skyplane.replicate.replication_plan import ReplicationTopology, ReplicationTopologyGateway
 from skyplane.utils import logger
 from skyplane.utils.definitions import gateway_docker_image
 from skyplane.utils.fn import PathLike, do_parallel
 
 if TYPE_CHECKING:
-    from skyplane.api.provision.provisioner import Provisioner
+    from skyplane.api.provisioner import Provisioner
 
 
 class DataplaneAutoDeprovision:
@@ -224,7 +223,7 @@ class Dataplane:
         self.jobs_to_dispatch.append(job)
         return job.uuid
 
-    def run_async(self, progress_reporter: TransferHook = None) -> TransferProgressTracker:
+    def run_async(self, progress_reporter: Optional[TransferHook] = None) -> TransferProgressTracker:
         if not self.provisioned:
             logger.error("Dataplane must be pre-provisioned. Call dataplane.provision() before starting a transfer")
         tracker = TransferProgressTracker(self, self.jobs_to_dispatch, self.transfer_config, progress_reporter)
@@ -234,7 +233,7 @@ class Dataplane:
         self.jobs_to_dispatch = []
         return tracker
 
-    def run(self, progress_reporter: TransferHook = None):
+    def run(self, progress_reporter: Optional[TransferHook] = None):
         tracker = self.run_async(progress_reporter)
         logger.fs.debug(f"[SkyplaneClient] Waiting for transfer to complete")
         tracker.join()
